use i64 in text_key_chain instead of i32
this fixes the fucking problem that was in the C version - unicode chars. i am using chars for u32 instead of as_bytes for u8, now. yippee

+++ src/core/misc.rs
@@ -71,7 +71,7 @@ impl EncryptedData
     }
 
     //CREATE Data WITH Success ExitCode
-    pub fn from(output: Vec<i32>, key: String) -> Self
+    pub fn from(output: Vec<i64>, key: String) -> Self
     {
         Self
         {
@@ -151,7 +151,7 @@ pub fn generate_key(length: usize) -> String
     Alphanumeric.sample_string(&mut rand::rng(), length)
 }
 
-pub fn generate_text_key_chain(key: &str, size: usize) -> Vec<i32>
+pub fn generate_text_key_chain(key: &str, size: usize) -> Vec<i64>
 {
     //VARIABLES
     let mut number_buffer: usize;
@@ -159,7 +159,7 @@ pub fn generate_text_key_chain(key: &str, size: usize) -> Vec<i32>
     let mut number_buffer_3: usize;
     let core_options = options::get_core_options();
     let key_length = core_options.key_length;
-    let mut text_key_chain: Vec<i32> = vec![0; size];
+    let mut text_key_chain: Vec<i64> = vec![0; size];
     let key_bytes = key.as_bytes();
 
     for i in 0..size
@@ -195,8 +195,8 @@ pub fn generate_text_key_chain(key: &str, size: usize) -> Vec<i32>
         }
 
         //VALUES
-        let a = key_bytes[number_buffer] as i32;
-        let b = key_bytes[number_buffer_3] as i32;
+        let a = key_bytes[number_buffer] as i64;
+        let b = key_bytes[number_buffer_3] as i64;
 
         //GET MATCHING OPERATION BETWEEN VALUES
         let val = if core_options.version == Version::V4 && (number_buffer + 1) % 4 == 0
